backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.4

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI and LangChain
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-openai==0.0.2
google-generativeai==0.3.2
openai==1.7.1
tiktoken==0.5.2

# Vector Database
qdrant-client==1.7.0

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
unstructured==0.11.8

# Redis
redis==5.0.1
aioredis==2.0.1

# AWS S3
boto3==1.34.14

# HTTP
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Monitoring
sentry-sdk[fastapi]==1.39.1
